requirements.txt
# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
neo4j==5.14.0
redis==5.0.1
alembic==1.12.1

# Scraping & OSINT
scrapy==2.11.0
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
aiohttp==3.9.0
playwright==1.40.0

# OSINT APIs
shodan==1.31.0
python-whois==0.8.0
dnspython==2.4.2
censys==2.2.8

# AI/ML
torch==2.1.0
transformers==4.35.0
sentence-transformers==2.2.2
spacy==3.7.2
scikit-learn==1.3.2
networkx==3.2.1
torch-geometric==2.4.0
datasets==2.15.0

# Tasks & Queue
celery==5.3.4
flower==2.0.1

# Utils
python-dotenv==1.0.0
loguru==0.7.2
pydantic==2.5.0
pytest==7.4.3
python-multipart==0.0.6

# Export & Reporting
reportlab==4.0.7
jinja2==3.1.2
matplotlib==3.8.2
seaborn==0.13.0

# Data augmentation
nlpaug==1.1.11

# Monitoring
prometheus-client==0.19.0
